+++ rcman-derive/tests/ui/fail/invalid_enum.rs
@@ -1,0 +1,16 @@
+use rcman_derive::SettingsSchema;
+use serde::{Deserialize, Serialize};
+
+#[derive(SettingsSchema, Serialize, Deserialize)]
+#[serde(tag = "type")]
+enum UnitVariantEnum {
+    Unit,
+}
+
+#[derive(SettingsSchema, Serialize, Deserialize)]
+#[serde(tag = "type")]
+enum MultiTupleEnum {
+    Multi(u32, String),
+}
+
+fn main() {}